timed searches for ffm

+++ chess4j-java/src/main/java/dev/jamesswafford/chess4j/search/SearchIterator.java
@@ -13,7 +13,7 @@ public interface SearchIterator {
 
     void setMaxDepth(int maxDepth);
 
-    void setMaxTime(long maxTimeMs);
+    void setMaxTime(int maxTimeMs);
 
     void setPost(boolean post);
